Safari-609.1.11.1/LayoutTests/fast/selectors/nth-child-of-chained-expected.txt

Test chaining many :nth-child() selectors. This verifies register allocation is correct.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing ":nth-child(-n+21 of testcase):nth-child(odd of testcase):nth-child(-2n+21 of testcase):nth-child(3n+1 of testcase):nth-child(-3n+22 of testcase):nth-child(4n+1 of testcase):nth-child(-5n+51 of testcase)"
PASS document.querySelectorAll('testcase:nth-child(-n+21 of testcase):nth-child(odd of testcase):nth-child(-2n+21 of testcase):nth-child(3n+1 of testcase):nth-child(-3n+22 of testcase):nth-child(4n+1 of testcase):nth-child(-5n+51 of testcase)').length is 1
PASS document.querySelectorAll('testcase:nth-child(-n+21 of testcase):nth-child(odd of testcase):nth-child(-2n+21 of testcase):nth-child(3n+1 of testcase):nth-child(-3n+22 of testcase):nth-child(4n+1 of testcase):nth-child(-5n+51 of testcase)')[0].id is "testcase1"
PASS getComputedStyle(document.querySelectorAll("testcase")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("testcase")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[4]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[5]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[6]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[7]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[8]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[9]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[10]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[11]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[12]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[13]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[14]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[15]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[16]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[17]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[18]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[19]).backgroundColor is "rgb(255, 0, 0)"
PASS document.querySelectorAll('div>testcase:nth-child(-n+21 of testcase):nth-child(odd of testcase):nth-child(-2n+21 of testcase):nth-child(3n+1 of testcase):nth-child(-3n+22 of testcase):nth-child(4n+1 of testcase):nth-child(-5n+51 of testcase)>div>div>div span').length is 1

Testing ":nth-child(2n-200 of testcase):nth-child(3n-1 of testcase):nth-child(4n-38 of testcase):nth-child(5n-98 of testcase)"
PASS document.querySelectorAll('testcase:nth-child(2n-200 of testcase):nth-child(3n-1 of testcase):nth-child(4n-38 of testcase):nth-child(5n-98 of testcase)').length is 1
PASS document.querySelectorAll('testcase:nth-child(2n-200 of testcase):nth-child(3n-1 of testcase):nth-child(4n-38 of testcase):nth-child(5n-98 of testcase)')[0].id is "testcase2"
PASS getComputedStyle(document.querySelectorAll("testcase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("testcase")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[4]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[5]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[6]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[7]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[8]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[9]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[10]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[11]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[12]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[13]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[14]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[15]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[16]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[17]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[18]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[19]).backgroundColor is "rgb(255, 0, 0)"
PASS document.querySelectorAll('div>testcase:nth-child(2n-200 of testcase):nth-child(3n-1 of testcase):nth-child(4n-38 of testcase):nth-child(5n-98 of testcase)>div>div>div span').length is 1

Testing ":nth-child(even of testcase):nth-child(3n of testcase):nth-child(4n-2 of testcase):nth-child(5n+1 of testcase)"
PASS document.querySelectorAll('testcase:nth-child(even of testcase):nth-child(3n of testcase):nth-child(4n-2 of testcase):nth-child(5n+1 of testcase)').length is 1
PASS document.querySelectorAll('testcase:nth-child(even of testcase):nth-child(3n of testcase):nth-child(4n-2 of testcase):nth-child(5n+1 of testcase)')[0].id is "testcase6"
PASS getComputedStyle(document.querySelectorAll("testcase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[4]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[5]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("testcase")[6]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[7]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[8]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[9]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[10]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[11]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[12]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[13]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[14]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[15]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[16]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[17]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[18]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[19]).backgroundColor is "rgb(255, 0, 0)"
PASS document.querySelectorAll('div>testcase:nth-child(even of testcase):nth-child(3n of testcase):nth-child(4n-2 of testcase):nth-child(5n+1 of testcase)>div>div>div span').length is 1

Testing ":nth-child(-n+9000 of testcase):nth-child(-2n+683 of testcase):nth-child(-3n+31 of testcase):nth-child(-4n+47 of testcase):nth-child(-5n+107 of testcase):nth-child(-6n+73 of testcase):nth-child(-7n+70 of testcase)"
PASS document.querySelectorAll('testcase:nth-child(-n+9000 of testcase):nth-child(-2n+683 of testcase):nth-child(-3n+31 of testcase):nth-child(-4n+47 of testcase):nth-child(-5n+107 of testcase):nth-child(-6n+73 of testcase):nth-child(-7n+70 of testcase)').length is 1
PASS document.querySelectorAll('testcase:nth-child(-n+9000 of testcase):nth-child(-2n+683 of testcase):nth-child(-3n+31 of testcase):nth-child(-4n+47 of testcase):nth-child(-5n+107 of testcase):nth-child(-6n+73 of testcase):nth-child(-7n+70 of testcase)')[0].id is "testcase7"
PASS getComputedStyle(document.querySelectorAll("testcase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[4]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[5]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[6]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("testcase")[7]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[8]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[9]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[10]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[11]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[12]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[13]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[14]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[15]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[16]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[17]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[18]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[19]).backgroundColor is "rgb(255, 0, 0)"
PASS document.querySelectorAll('div>testcase:nth-child(-n+9000 of testcase):nth-child(-2n+683 of testcase):nth-child(-3n+31 of testcase):nth-child(-4n+47 of testcase):nth-child(-5n+107 of testcase):nth-child(-6n+73 of testcase):nth-child(-7n+70 of testcase)>div>div>div span').length is 1

Testing ":nth-child(odd of testcase):nth-child(3n+1 of testcase):nth-child(4n-1 of testcase):nth-child(5n+4 of testcase)"
PASS document.querySelectorAll('testcase:nth-child(odd of testcase):nth-child(3n+1 of testcase):nth-child(4n-1 of testcase):nth-child(5n+4 of testcase)').length is 1
PASS document.querySelectorAll('testcase:nth-child(odd of testcase):nth-child(3n+1 of testcase):nth-child(4n-1 of testcase):nth-child(5n+4 of testcase)')[0].id is "testcase19"
PASS getComputedStyle(document.querySelectorAll("testcase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[4]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[5]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[6]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[7]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[8]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[9]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[10]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[11]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[12]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[13]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[14]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[15]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[16]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[17]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[18]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("testcase")[19]).backgroundColor is "rgb(255, 0, 0)"
PASS document.querySelectorAll('div>testcase:nth-child(odd of testcase):nth-child(3n+1 of testcase):nth-child(4n-1 of testcase):nth-child(5n+4 of testcase)>div>div>div span').length is 1

Testing ":nth-child(-n+1500 of testcase):nth-child(2n+1 of testcase):nth-child(-3n+45 of testcase):nth-child(4n+3 of testcase):nth-child(-5n+1545 of testcase)"
PASS document.querySelectorAll('testcase:nth-child(-n+1500 of testcase):nth-child(2n+1 of testcase):nth-child(-3n+45 of testcase):nth-child(4n+3 of testcase):nth-child(-5n+1545 of testcase)').length is 1
PASS document.querySelectorAll('testcase:nth-child(-n+1500 of testcase):nth-child(2n+1 of testcase):nth-child(-3n+45 of testcase):nth-child(4n+3 of testcase):nth-child(-5n+1545 of testcase)')[0].id is "testcase15"
PASS getComputedStyle(document.querySelectorAll("testcase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[4]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[5]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[6]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[7]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[8]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[9]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[10]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[11]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[12]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[13]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[14]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("testcase")[15]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[16]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[17]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[18]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[19]).backgroundColor is "rgb(255, 0, 0)"
PASS document.querySelectorAll('div>testcase:nth-child(-n+1500 of testcase):nth-child(2n+1 of testcase):nth-child(-3n+45 of testcase):nth-child(4n+3 of testcase):nth-child(-5n+1545 of testcase)>div>div>div span').length is 1

Testing ":nth-child(n+2 of testcase):nth-child(-n+6000 of testcase):nth-child(n-2 of testcase):nth-child(2n of testcase):nth-child(-2n+20 of testcase):nth-child(2n+2 of testcase):nth-child(3n+1 of testcase):nth-child(-3n+49 of testcase):nth-child(3n-92 of testcase):nth-child(4n+2 of testcase):nth-child(-4n+50 of testcase)"
PASS document.querySelectorAll('testcase:nth-child(n+2 of testcase):nth-child(-n+6000 of testcase):nth-child(n-2 of testcase):nth-child(2n of testcase):nth-child(-2n+20 of testcase):nth-child(2n+2 of testcase):nth-child(3n+1 of testcase):nth-child(-3n+49 of testcase):nth-child(3n-92 of testcase):nth-child(4n+2 of testcase):nth-child(-4n+50 of testcase)').length is 1
PASS document.querySelectorAll('testcase:nth-child(n+2 of testcase):nth-child(-n+6000 of testcase):nth-child(n-2 of testcase):nth-child(2n of testcase):nth-child(-2n+20 of testcase):nth-child(2n+2 of testcase):nth-child(3n+1 of testcase):nth-child(-3n+49 of testcase):nth-child(3n-92 of testcase):nth-child(4n+2 of testcase):nth-child(-4n+50 of testcase)')[0].id is "testcase10"
PASS getComputedStyle(document.querySelectorAll("testcase")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[4]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[5]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[6]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[7]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[8]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[9]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("testcase")[10]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[11]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[12]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[13]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[14]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[15]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[16]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[17]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[18]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("testcase")[19]).backgroundColor is "rgb(255, 0, 0)"
PASS document.querySelectorAll('div>testcase:nth-child(n+2 of testcase):nth-child(-n+6000 of testcase):nth-child(n-2 of testcase):nth-child(2n of testcase):nth-child(-2n+20 of testcase):nth-child(2n+2 of testcase):nth-child(3n+1 of testcase):nth-child(-3n+49 of testcase):nth-child(3n-92 of testcase):nth-child(4n+2 of testcase):nth-child(-4n+50 of testcase)>div>div>div span').length is 1

PASS successfullyParsed is true

TEST COMPLETE
